fix: make response mapper a handler method and decouple update request
