+++ internal/guest/handler/customer/get_me_test.go
@@ -20,18 +20,20 @@ func TestGetMe(t *testing.T) {
 		userID     = gofakeit.UUID()
 		customerID = gofakeit.UUID()
 
-		userName        = "sharebite04"
-		firstName       = gofakeit.FirstName()
-		lastName        = gofakeit.LastName()
+		userName  = "sharebite04"
+		firstName = gofakeit.FirstName()
+		lastName  = gofakeit.LastName()
+		bio       = gofakeit.Bio()
+
+		baseURL         = "https://cdn.com/"
 		avatarObjectKey = gofakeit.Word()
-		bio             = gofakeit.Bio()
 	)
 
 	tests := []struct {
 		name string
 
 		userID any
-		mockFn func(s *mockCustomerService)
+		mockFn func(s *mockCustomerService, st *mockObjectStorage)
 
 		wantBody any
 		wantCode int
@@ -39,7 +41,7 @@ func TestGetMe(t *testing.T) {
 		{
 			name:   "success",
 			userID: userID,
-			mockFn: func(s *mockCustomerService) {
+			mockFn: func(s *mockCustomerService, st *mockObjectStorage) {
 				s.On("GetByUserID", mock.Anything, userID).
 					Return(entity.Customer{
 						ID:              customerID,
@@ -51,24 +53,28 @@ func TestGetMe(t *testing.T) {
 						Bio:             &bio,
 					}, nil).
 					Once()
+
+				st.On("BuildURL", avatarObjectKey).
+					Return(baseURL + avatarObjectKey).
+					Once()
 			},
 			wantBody: getMeResponse{
-				Customer: customerToResponse(entity.Customer{
-					ID:              customerID,
-					UserID:          userID,
-					UserName:        userName,
-					FirstName:       firstName,
-					LastName:        lastName,
-					AvatarObjectKey: &avatarObjectKey,
-					Bio:             &bio,
-				}),
+				Customer: customerResponse{
+					ID:        customerID,
+					UserID:    userID,
+					UserName:  userName,
+					FirstName: firstName,
+					LastName:  lastName,
+					Bio:       &bio,
+					AvatarURL: strPtr(baseURL + avatarObjectKey),
+				},
 			},
 			wantCode: http.StatusOK,
 		},
 		{
 			name:   "error - customer not found",
 			userID: userID,
-			mockFn: func(s *mockCustomerService) {
+			mockFn: func(s *mockCustomerService, st *mockObjectStorage) {
 				s.On("GetByUserID", mock.Anything, userID).
 					Return(entity.Customer{}, apperror.CustomerNotFoundUserID(userID)).
 					Once()
@@ -79,14 +85,14 @@ func TestGetMe(t *testing.T) {
 		{
 			name:     "unauthorized - no user id in context",
 			userID:   nil,
-			mockFn:   func(s *mockCustomerService) {},
+			mockFn:   func(s *mockCustomerService, st *mockObjectStorage) {},
 			wantCode: http.StatusInternalServerError, // we expect to have token in context after auth middleware step!
 			wantBody: response.ErrorResponse{Message: internalErrMsg},
 		},
 		{
 			name:   "service unknown error",
 			userID: userID,
-			mockFn: func(s *mockCustomerService) {
+			mockFn: func(s *mockCustomerService, st *mockObjectStorage) {
 				s.On("GetByUserID", mock.Anything, userID).
 					Return(entity.Customer{}, errors.New("database connection lost")).
 					Once()
@@ -101,8 +107,9 @@ func TestGetMe(t *testing.T) {
 			t.Parallel()
 
 			customerService := new(mockCustomerService)
-			h := &handler{service: customerService}
-			tt.mockFn(customerService)
+			storage := new(mockObjectStorage)
+			h := &handler{service: customerService, storage: storage}
+			tt.mockFn(customerService, storage)
 
 			r := newTestRouter()
 			r.GET("/customers", withUserID(tt.userID, h.getMe))
@@ -114,6 +121,7 @@ func TestGetMe(t *testing.T) {
 			}
 
 			customerService.AssertExpectations(t)
+			storage.AssertExpectations(t)
 		})
 	}
 }
